// lionpay-management-old/src/generated-api/wallet/api.ts
export interface AdjustBalanceRequestAmount {
}
// ...
export interface TransactionResponse {
    'txId': string;
    'merchantId': string;
    'amount': AdjustBalanceRequestAmount;
    'txType': TxType;
    'txStatus': TxStatus;
    'merchantName': string;
    'balanceAfter': AdjustBalanceRequestAmount;
    'currency': string;
    'originalAmount': AdjustBalanceRequestAmount;
    'createdAt': string;
}



export const TxStatus = {
    Success: 'Success',
    Failed: 'Failed'
} as const;

export type TxStatus = typeof TxStatus[keyof typeof TxStatus];



export const TxType = {
    Payment: 'Payment',
    Charge: 'Charge',
    AdminCharge: 'AdminCharge',
    AdminDeduct: 'AdminDeduct'
} as const;

export type TxType = typeof TxType[keyof typeof TxType];